refactor: remove the v0.2.0 demo
Closes #304
