+++ src/wp-notifications.js
@@ -1,4 +1,3 @@
 import './scripts/';
-import './scripts/demo/';
 
 import './styles/wp-notifications.scss';
